#1521 - Created the contact line field
