+++ sormas-ui/src/main/java/de/symeda/sormas/ui/contact/components/linelisting/LineListingLayout.java
@@ -1,19 +1,16 @@
 package de.symeda.sormas.ui.contact.components.linelisting;
 
 import java.io.Serializable;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 import com.vaadin.data.Binder;
-import com.vaadin.data.BinderValidationStatus;
 import com.vaadin.icons.VaadinIcons;
 import com.vaadin.ui.Alignment;
 import com.vaadin.ui.Button;
 import com.vaadin.ui.ComboBox;
-import com.vaadin.ui.DateField;
 import com.vaadin.ui.HorizontalLayout;
 import com.vaadin.ui.Label;
 import com.vaadin.ui.VerticalLayout;
@@ -22,11 +19,7 @@ import com.vaadin.ui.themes.ValoTheme;
 
 import de.symeda.sormas.api.Disease;
 import de.symeda.sormas.api.FacadeProvider;
-import de.symeda.sormas.api.caze.CaseDataDto;
 import de.symeda.sormas.api.caze.CaseReferenceDto;
-import de.symeda.sormas.api.contact.ContactDto;
-import de.symeda.sormas.api.contact.ContactProximity;
-import de.symeda.sormas.api.contact.ContactRelation;
 import de.symeda.sormas.api.i18n.Captions;
 import de.symeda.sormas.api.i18n.I18nProperties;
 import de.symeda.sormas.api.i18n.Strings;
@@ -35,14 +28,13 @@ import de.symeda.sormas.api.region.RegionReferenceDto;
 import de.symeda.sormas.api.user.UserRole;
 import de.symeda.sormas.api.utils.ValidationRuntimeException;
 import de.symeda.sormas.ui.UserProvider;
+import de.symeda.sormas.ui.contact.components.linelisting.contactfield.ContactFieldDto;
+import de.symeda.sormas.ui.contact.components.linelisting.contactfield.ContactLineField;
 import de.symeda.sormas.ui.utils.ButtonHelper;
 import de.symeda.sormas.ui.utils.CssStyles;
 import de.symeda.sormas.ui.utils.FieldHelper;
 import de.symeda.sormas.ui.utils.components.linelisting.line.DeleteLineEvent;
 import de.symeda.sormas.ui.utils.components.linelisting.line.LineLayout;
-import de.symeda.sormas.ui.utils.components.linelisting.person.PersonField;
-import de.symeda.sormas.ui.utils.components.linelisting.person.PersonFieldDto;
-import de.symeda.sormas.ui.utils.components.multidayselector.MultiDaySelectorDto;
 
 public class LineListingLayout extends VerticalLayout {
 
@@ -202,10 +194,7 @@ public class LineListingLayout extends VerticalLayout {
 			newLineDto.setDisease(lastLineDto.getDisease());
 			newLineDto.setRegion(lastLineDto.getRegion());
 			newLineDto.setDistrict(lastLineDto.getDistrict());
-			newLineDto.setDateOfReport(lastLineDto.getDateOfReport());
-			newLineDto.setMultiDaySelector(lastLineDto.getMultiDaySelector());
-			newLineDto.setTypeOfContact(lastLineDto.getTypeOfContact());
-			newLineDto.setRelationToCase(lastLineDto.getRelationToCase());
+			newLineDto.setLineField(lastLineDto.getLineField());
 		}
 
 		newLine.setBean(newLineDto);
@@ -226,13 +215,7 @@ public class LineListingLayout extends VerticalLayout {
 
 		private final Binder<ContactLineDto> binder = new Binder<>(ContactLineDto.class);
 
-		private final DateField dateOfReport;
-		private final MultiDayContactField multiDay;
-
-		private final ComboBox<ContactProximity> typeOfContact;
-		private final ComboBox<ContactRelation> relationToCase;
-		private final PersonField person;
-
+		private final ContactLineField contactLineField;
 		private final Button delete;
 
 		public ContactLineLayout(int lineIndex) {
@@ -245,38 +228,19 @@ public class LineListingLayout extends VerticalLayout {
 			binder.forField(region).asRequired().bind(ContactLineDto.REGION);
 			binder.forField(district).asRequired().bind(ContactLineDto.DISTRICT);
 
-			dateOfReport = new DateField();
-			dateOfReport.setId("lineListingDateOfReport_" + lineIndex);
-			dateOfReport.setWidth(150, Unit.PIXELS);
-			binder.forField(dateOfReport).asRequired().bind(ContactLineDto.DATE_OF_REPORT);
-			dateOfReport.setRangeEnd(LocalDate.now());
-
-			multiDay = new MultiDayContactField();
-			multiDay.setId("lineListingMultiDay_" + lineIndex);
-			binder.forField(multiDay).bind(ContactLineDto.MULTI_DAY_SELECTOR);
-
-			typeOfContact = new ComboBox<>();
-			typeOfContact.setId("lineListingContactProximity_" + lineIndex);
-			typeOfContact.setItems(ContactProximity.values());
-			typeOfContact.setWidth(150, Unit.PIXELS);
-			typeOfContact.addStyleName(CssStyles.CAPTION_OVERFLOW);
-			binder.forField(typeOfContact).bind(ContactLineDto.TYPE_OF_CONTACT);
-
-			relationToCase = new ComboBox<>();
-			relationToCase.setId("lineListingRelationToCase_" + lineIndex);
-			relationToCase.setItems(ContactRelation.values());
-			relationToCase.setWidth(150, Unit.PIXELS);
-			relationToCase.addStyleName(CssStyles.CAPTION_OVERFLOW);
-			binder.forField(relationToCase).bind(ContactLineDto.RELATION_TO_CASE);
-
-			person = new PersonField();
-			person.setId("lineListingPerson_" + lineIndex);
-			binder.forField(person).bind(ContactLineDto.PERSON);
+			contactLineField = new ContactLineField();
+			contactLineField.setId("lineListingContactLineField_" + lineIndex);
+			binder.forField(contactLineField).bind(ContactLineDto.LINE_FIELD);
 
 			delete = ButtonHelper
 				.createIconButtonWithCaption("delete_" + lineIndex, null, VaadinIcons.TRASH, event -> fireEvent(new DeleteLineEvent(this)));
+			delete.setStyleName(CssStyles.VSPACE_3);
 
-			addComponents(dateOfReport, multiDay, typeOfContact, relationToCase, person, delete);
+			addComponents(contactLineField, delete);
+
+			setComponentAlignment(contactLineField, Alignment.BOTTOM_LEFT);
+			setComponentAlignment(delete, Alignment.BOTTOM_LEFT);
+
 			formatLine(lineIndex);
 		}
 
@@ -289,28 +253,13 @@ public class LineListingLayout extends VerticalLayout {
 		}
 
 		public boolean hasErrors() {
-			BinderValidationStatus<MultiDaySelectorDto> contactDatesStatus = multiDay.validate();
-			BinderValidationStatus<PersonFieldDto> personValidationStatus = person.validate();
-			BinderValidationStatus<ContactLineDto> lineValidationStatus = binder.validate();
-			return contactDatesStatus.hasErrors() || personValidationStatus.hasErrors() || lineValidationStatus.hasErrors();
+			return contactLineField.hasErrors();
 		}
 
 		private void formatLine(int lineIndex) {
 
-			dateOfReport.setCaption(I18nProperties.getPrefixCaption(CaseDataDto.I18N_PREFIX, CaseDataDto.REPORT_DATE));
-			dateOfReport.removeStyleName(CssStyles.CAPTION_HIDDEN);
-			setComponentAlignment(dateOfReport, Alignment.BOTTOM_LEFT);
-			multiDay.showCaptions();
-			typeOfContact.setCaption(I18nProperties.getPrefixCaption(ContactDto.I18N_PREFIX, ContactDto.CONTACT_PROXIMITY));
-			typeOfContact.removeStyleName(CssStyles.CAPTION_HIDDEN);
-			setComponentAlignment(typeOfContact, Alignment.BOTTOM_LEFT);
-			relationToCase.setCaption(I18nProperties.getPrefixCaption(ContactDto.I18N_PREFIX, ContactDto.RELATION_TO_CASE));
-			relationToCase.removeStyleName(CssStyles.CAPTION_HIDDEN);
-			setComponentAlignment(relationToCase, Alignment.BOTTOM_LEFT);
-			person.showCaptions();
-			setComponentAlignment(person, Alignment.BOTTOM_LEFT);
+			contactLineField.showCaptions();
 			delete.setEnabled(lineIndex > 0);
-			setComponentAlignment(delete, Alignment.MIDDLE_LEFT);
 		}
 
 		public Button getDelete() {
@@ -324,21 +273,13 @@ public class LineListingLayout extends VerticalLayout {
 		public static final String DISEASE = "disease";
 		public static final String REGION = "region";
 		public static final String DISTRICT = "district";
-		public static final String DATE_OF_REPORT = "dateOfReport";
-		public static final String MULTI_DAY_SELECTOR = "multiDaySelector";
-		public static final String TYPE_OF_CONTACT = "typeOfContact";
-		public static final String RELATION_TO_CASE = "relationToCase";
-		public static final String PERSON = "person";
+		public static final String LINE_FIELD = "lineField";
 
 		private CaseReferenceDto caze;
 		private Disease disease;
 		private RegionReferenceDto region;
 		private DistrictReferenceDto district;
-		private LocalDate dateOfReport;
-		private MultiDaySelectorDto multiDaySelector;
-		private ContactProximity typeOfContact;
-		private ContactRelation relationToCase;
-		private PersonFieldDto person;
+		private ContactFieldDto lineField;
 
 		public CaseReferenceDto getCaze() {
 			return caze;
@@ -372,44 +313,12 @@ public class LineListingLayout extends VerticalLayout {
 			this.district = district;
 		}
 
-		public LocalDate getDateOfReport() {
-			return dateOfReport;
+		public ContactFieldDto getLineField() {
+			return lineField;
 		}
 
-		public void setDateOfReport(LocalDate dateOfReport) {
-			this.dateOfReport = dateOfReport;
-		}
-
-		public MultiDaySelectorDto getMultiDaySelector() {
-			return multiDaySelector;
-		}
-
-		public void setMultiDaySelector(MultiDaySelectorDto multiDaySelector) {
-			this.multiDaySelector = multiDaySelector;
-		}
-
-		public ContactProximity getTypeOfContact() {
-			return typeOfContact;
-		}
-
-		public void setTypeOfContact(ContactProximity typeOfContact) {
-			this.typeOfContact = typeOfContact;
-		}
-
-		public ContactRelation getRelationToCase() {
-			return relationToCase;
-		}
-
-		public void setRelationToCase(ContactRelation relationToCase) {
-			this.relationToCase = relationToCase;
-		}
-
-		public PersonFieldDto getPerson() {
-			return person;
-		}
-
-		public void setPerson(PersonFieldDto person) {
-			this.person = person;
+		public void setLineField(ContactFieldDto lineField) {
+			this.lineField = lineField;
 		}
 	}
 }
